debug feeature: improve heap debugging on MSVC using the compiler-provided MSVCRT/CRTDBG plus a bit of juice of our own.
NOTE: those markers MUST follow the #pragma init_seg(lib) or it won't fly they MUST be included in the same segment as the globals they're 'wrapping'.

The markers are defined in the external mupdf monolithic library code. See the GerHobbelt/mupdf.git project. When this library does not define HAVE_MUPDF a default (nil) implementation is used instead.